More texture stuff
Minor cleanup in Env Map test.  Pass placeholder image as filename directly.  Materials omit overriding to black (the default).

Removed need for idiot alarm to be bypassed for Shape overwrites - now the base class just doesn't trigger it if the context is being re-used
  Does this still leave an opening to call copy-to-gpu too much?
    Is there any use case where the user might call copy-to-gpu every frame on an existing shape made in the constructor?
      Probably not.  We're good.

Materials don't try to activate textures that aren't ready, which would otherwise turn textures on when nothing is there to load.