Remove extra functions from core
GPU object no longer needs an extra abstract function for making blank gpu representations - that is all handled in an extra argument to copy-to-gpu now.  The other arguments turned out to be unused and were removed.  The check_idiot_alarm function no longer needs to be overridden and was thus inlined away.