Eliminated class Material and interface Container.
Two classes no longer exist or need to be learned about, yet using materials requires no significant increase in typing with the change, and far easier to grasp.

Instead of having override() be an obtuse part of the API to learn, declaring it right before each usage in each demo, as a self-evident one-liner shorthand hack.  This lets us be a little more situation-specific too, such as only overriding colors for the colors tutorial.

In the next commit, even this helper function will be eliminated -- since es6 syntax DOES have a way of overriding properties afterall: { ...original, new_property = a }.